backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
openai==1.55.3
gunicorn==20.1.0
httpx==0.27.2
